+++ src/compas_invocations2/grasshopper.py
@@ -1,8 +1,11 @@
 import os
 import shutil
+import tempfile
 
 import invoke
 import requests
+
+from compas_invocations2.console import chdir
 
 YAK_URL = r"https://files.mcneel.com/yak/tools/latest/yak.exe"
 FILENAME = "yak.exe"
@@ -117,3 +120,83 @@ def yakerize(
     if not any([f.endswith(".yak") for f in os.listdir(target_dir)]):
         invoke.Exit("No .yak file was created in the build directory.")
     os.chdir(path_current)
+
+
+@invoke.task(
+    help={
+        "gh_io_folder": "Folder where GH_IO.dll is located. If not specified, it will try to download from NuGet.",
+        "ironpython": "Command for running the IronPython executable. Defaults to `ipy`.",
+        "prefix": "(Optional) Append this prefix to the names of the built components.",
+    }
+)
+def build_ghuser_components(ctx, gh_io_folder=None, ironpython=None, prefix=None):
+    """Builds Grasshopper components using GH Componentizer."""
+    prefix = prefix or getattr(ctx.ghuser, "prefix", None)
+    source_dir = os.path.abspath(ctx.ghuser.source_dir)
+    target_dir = os.path.abspath(ctx.ghuser.target_dir)
+    repo_url = "https://github.com/compas-dev/compas-actions.ghpython_components.git"
+
+    with chdir(ctx.base_folder):
+        shutil.rmtree(os.path.join(ctx.base_folder, target_dir), ignore_errors=True)
+
+    # Build IronPython Grasshopper user objects from source
+    with chdir(ctx.base_folder):
+        with tempfile.TemporaryDirectory("actions.ghcomponentizer") as action_dir:
+            ctx.run("git clone {} {}".format(repo_url, action_dir))
+
+            if not gh_io_folder:
+                gh_io_folder = tempfile.mkdtemp("ghio")
+                import compas_ghpython
+
+                compas_ghpython.fetch_ghio_lib(gh_io_folder)
+
+            if not ironpython:
+                ironpython = ctx.get("ironpython") or "ipy"
+
+            gh_io_folder = os.path.abspath(gh_io_folder)
+            componentizer_script = os.path.join(action_dir, "componentize_ipy.py")
+
+            cmd = "{} {} {} {}".format(ironpython, componentizer_script, source_dir, target_dir)
+            cmd += ' --ghio "{}"'.format(gh_io_folder)
+            if prefix:
+                cmd += ' --prefix "{}"'.format(prefix)
+
+            ctx.run(cmd)
+
+
+@invoke.task(
+    help={
+        "gh_io_folder": "Folder where GH_IO.dll is located. If not specified, it will try to download from NuGet.",
+        "prefix": "(Optional) Append this prefix to the names of the built components.",
+    }
+)
+def build_cpython_ghuser_components(ctx, gh_io_folder=None, prefix=None):
+    """Builds CPython Grasshopper components using GH Componentizer."""
+    prefix = prefix or getattr(ctx.ghuser_cpython, "prefix", None)
+    source_dir = os.path.abspath(ctx.ghuser_cpython.source_dir)
+    target_dir = os.path.abspath(ctx.ghuser_cpython.target_dir)
+    repo_url = "https://github.com/compas-dev/compas-actions.ghpython_components.git"
+
+    with chdir(ctx.base_folder):
+        shutil.rmtree(os.path.join(ctx.base_folder, target_dir), ignore_errors=True)
+
+    # Build CPython Grasshopper user objects from source
+    with chdir(ctx.base_folder):
+        with tempfile.TemporaryDirectory("actions.ghcomponentizer") as action_dir:
+            ctx.run("git clone {} {}".format(repo_url, action_dir))
+
+            if not gh_io_folder:
+                gh_io_folder = tempfile.mkdtemp("ghio")
+                import compas_ghpython
+
+                compas_ghpython.fetch_ghio_lib(gh_io_folder)
+
+            gh_io_folder = os.path.abspath(gh_io_folder)
+            componentizer_script = os.path.join(action_dir, "componentize_cpy.py")
+
+            cmd = "{} {} {} {}".format("python", componentizer_script, source_dir, target_dir)
+            cmd += ' --ghio "{}"'.format(gh_io_folder)
+            if prefix:
+                cmd += ' --prefix "{}"'.format(prefix)
+
+            ctx.run(cmd)
